requirements.txt
# 🔌 WebSocket и HTTP клиенты
websockets>=12.0
aiohttp>=3.8.0

# 🗄️ PostgreSQL dependencies
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# 📊 Обработка данных
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# 🌐 API framework  
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0

# 📈 Monitoring and metrics
prometheus-client>=0.19.0
psutil>=5.9.0

# ⚙️ Configuration and environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# 📝 Logging
structlog>=23.0.0
colorlog>=6.7.0

# 🔐 Security
cryptography>=41.0.0

# 🧪 Тестирование
pytest>=7.0.0
pytest-asyncio>=0.21.0

# 📈 Мониторинг (опционально)
flask>=2.3.0
streamlit>=1.25.0

# 🛠️ Утилиты
python-dotenv>=1.0.0
pydantic>=2.0.0